PeopleDetection/requirements.txt
torch~=2.0.1
transformers~=4.30.2
torchvision~=0.15.2
requests~=2.31.0
opencv-python~=4.8.0.74
Pillow~=10.0.0
timm~=0.9.2